feat: Implement core Vaultis contract logic

+++ contracts/Counter.sol
@@ -2,18 +2,18 @@
 pragma solidity ^0.8.28;
 
 contract Counter {
-  uint public x;
+    uint256 public x;
 
-  event Increment(uint by);
+    event Increment(uint256 by);
 
-  function inc() public {
-    x++;
-    emit Increment(1);
-  }
+    function inc() public {
+        x++;
+        emit Increment(1);
+    }
 
-  function incBy(uint by) public {
-    require(by > 0, "incBy: increment should be positive");
-    x += by;
-    emit Increment(by);
-  }
+    function incBy(uint256 by) public {
+        require(by > 0, "incBy: increment should be positive");
+        x += by;
+        emit Increment(by);
+    }
 }
